Sometimes the VM needs restarting even when it's supposedly running containerd.
The VM gets in a state where it thinks it's running containerd, but it isn't.
Restarting it fixes this.

I read through the code and didn't see anything that would put the VM in this state,
but there might be a cause of an instability bug here waiting to be figured out.

Signed-off-by: Eric Promislow <epromislow@suse.com>